add weekly report commands
